backend/requirements.txt
flask==3.0.0
pymongo==4.6.1
python-dotenv==1.0.0
flask-cors==4.0.0
python-jose==3.3.0
pytest==8.0.0
python-multipart==0.0.9
requests==2.31.0
google-generativeai==0.3.2
werkzeug==3.0.1
